Made somewhere to store the code that was output from the converter
This file stores the plaintext of the results of using Builder.io to convert the Figma design to React. Each page is separated with a header, and each file for each page is also separated with a header. Do not merge/pull this; it's merely being used as a reference.